requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
chromadb
langdetect
httpx
gspread
redis